refactor: split progress.py into live / summaries / pipeline_tracker
Break the 1200-line progress.py into three focused modules (code moved
verbatim): live.py (LiveTableMonitor/LiveProgressBar/SetupProgress + spinner/
bar/CPU-RAM helpers), summaries.py (per-step summaries + detail blocks + final
panel), pipeline_tracker.py (PipelineTracker + tracker). Layering is acyclic
(live, summaries are leaves; tracker composes both; final_summary_panel imports
tracker lazily). progress.py now re-exports every public name, so all existing
'from ...progress import X' imports keep working. Verified: cli + all importers
load, every external name resolves, tracker is the same singleton.
